frontend/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.1
langchain-openai>=0.0.1
langsmith>=0.0.69
python-dotenv>=1.0.0
streamlit>=1.29.0
faiss-cpu>=1.7.4
torch>=2.1.0
pillow>=10.0.0
pydantic>=2.5.0
openai>=1.6.0
tiktoken>=0.5.2
